bot/pyproject.toml
[project]
name = "lab7-bot"
version = "0.1.0"
description = "Telegram bot scaffold for se-toolkit-lab-7"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "python-telegram-bot>=21.6",
]

[tool.uv]
package = false
